greengrass: Add core security resources and root key to image
Install the Greengrass core security resources and the root key
on the image. The user is expected to place a copy of
hash-setup.tar.gz and root.ca.pem in the root of the Yocto
Project environment (i.e., the folder containing the build
folder(s) and the sources folder) with the hash replaced with
'my-core'.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>